backend/requirements.txt
flask
flask-cors
python-dotenv
pillow
google-generativeai